Update aCamera.vs
Updated with support for aInventory library. When used together if the camera zooms, the aInventory outline filter will change its thickness so it looks better at each scale of the map. Without this the thickness would thin out at higher scales and looks awkward. This fixes that if the two libraries are used together.